change source dir name from "lib" to "src"
