+++ composer.json
@@ -18,7 +18,7 @@
   },
   "autoload": {
     "psr-4": {
-      "BotBoris\\": "lib/",
+      "BotBoris\\": "src/",
       "Tests\\": "tests/"
     }
   },
