+++ src/Event/HelloWorld.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace BotBoris\Event;
+
+class HelloWorld extends Event
+{
+    public function isConditionCompleted(): bool
+    {
+        return false;
+    }
+
+    public function getMessage(): string
+    {
+        return 'Hello, world!';
+    }
+}
