+++ src/Registry/Registry.php
